# api/apidog/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DAC",
    "description": "API do Back-end do Bantads",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/contas/{numero}/saldo": {
      "post": {
        "summary": "/contas/{numero}/saldo",
        "deprecated": false,
        "description": "retorna o saldo da conta",
        "tags": [
          "Contas"
        ],
        "parameters": [
          {
            "name": "numero",
            "in": "path",
            "description": "Número da conta",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Saldo retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaldoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Filtro Não Existe",
            "headers": {}
          },
          "401": {
            "description": "O usuário não está logado",
            "headers": {}
          },
          "403": {
            "description": "O usuário não tem permissão para efetuar esta operação",
            "headers": {}
          },
          "404": {
            "description": "Clientes não encontradas",
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/contas/{numero}/depositar": {
      "post": {
        "summary": "/contas/{numero}/depositar",
        "deprecated": false,
        "description": "deposita um valor na conta do cliente",
        "tags": [
          "Contas"
        ],
        "parameters": [
          {
            "name": "numero",
            "in": "path",
            "description": "Número da conta",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "valor": {
                    "type": "number",
                    "format": "double",
                    "description": "Valor a ser depositado",
                    "examples": [
                      51.44
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Depósito efetuado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperacaoResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "O usuário não está logado",
            "headers": {}
          },
          "403": {
            "description": "O usuário não tem permissão para efetuar esta operação",
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "SaldoResponse": {
        "type": "object",
        "properties": {
          "cliente": {
            "type": "string",
            "examples": [
              "76179646090"
            ]
          },
          "conta": {
            "type": "string",
            "examples": [
              "0556"
            ]
          },
          "saldo": {
            "type": "number",
            "format": "double",
            "examples": [
              1506.88
            ]
          }
        }
      },
...
      "OperacaoResponse": {
        "type": "object",
        "properties": {
          "conta": {
            "type": "string",
            "examples": [
              "1291"
            ]
          },
          "data": {
            "type": "string",
            "examples": [
              "2025-08-01T10:22:45-03:00"
            ]
          },
          "saldo": {
            "type": "number",
            "format": "double",
            "examples": [
              1044.23
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "JTW Token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Develop Env"
    },
    {
      "url": "https://prod.your-api-server.com",
      "description": "Prod Env"
    }
  ],
  "security": []
}